pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "app"
version = "0.0.1"
description = ""
authors = ["Sam Dunham <sdunham@nvidia.com>"]
include = ["build/*"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.111.0"
pydantic = "^2.8.2"
pydantic_settings = "^2.4.0"
uvicorn = "^0.18.1"
requests = "^2.25.0"
APScheduler = "^3.9.1"
httpx = "^0.23.0"
websockets = "^10.3"
storable = "^1.2.4"
PyYAML = "^6.0"
fastapi-azure-auth = "^3.5.1"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pytest = "^7.1.2"
flake8 = "^4.0.1"
pytest-asyncio = "^0.19.0"

[[tool.poetry.source]]
name = "nv-shared"
url = "https://urm.nvidia.com/artifactory/api/pypi/nv-shared-pypi/simple/"
priority = "explicit"

[tool.poetry.scripts]
start = "app.src.main:start"


[tool.isort]
profile = "black"
skip_gitignore = true

